+++ src/bot.ts
@@ -1,9 +1,9 @@
-import { Client, Collection } from "discord.js";
 import { REST } from "@discordjs/rest";
-import { Routes } from "discord-api-types/v9";
+import * as events from "@events/index";
 import Command from "@models/command";
 import LoadCommands from "@util/command_loader";
-import * as events from "@events/index";
+import { Routes } from "discord-api-types/v9";
+import { Client, Collection } from "discord.js";
 
 export class Bot {
     /** The discordjs client object */
@@ -41,8 +41,7 @@ export class Bot {
         // Load the commands into the bot
         LoadCommands().then(cmds => {
             this.commands = cmds;
-            this.registerSlashCommands(
-                new Array(...new Set(cmds.map(c => c.SlashCommand).filter(c => c !== undefined))));
+            this.registerSlashCommands(cmds.map(c => c.SlashCommand));
         });
     }
 
@@ -51,7 +50,6 @@ export class Bot {
      */
     public eventSubscriber(): void {
         this.client.once("ready", () => events.ready(this));
-        this.client.on("messageCreate", msg => events.messageCreate(msg, this));
         this.client.on("interactionCreate", interaction => events.interactionCreate(interaction, this));
     }
 
@@ -61,7 +59,6 @@ export class Bot {
     public async registerSlashCommands(cmds: object[]) {
         /// TODO: Check if production and change to applicationCommands. Change back to applicationGuildCommands for testing.
         /// TODO: Look into slash command permissions: https://discordjs.guide/interactions/slash-command-permissions.html#user-permissions
-        /// TODO: Migrate slash commands to a separate class in the same file so they can access the commands for things like help menu options
         await this.rest.put(Routes.applicationCommands(process.env.CLIENT_ID!), { body: cmds })
             .then(() => console.log(`Loaded ${cmds.length} slash commands.`))
             .catch(e => console.error(e));