+++ src/commands/mod/unban.ts
@@ -1,34 +1,15 @@
-import { CommandInteraction, Message } from "discord.js";
-import { SlashCommandBuilder } from "@discordjs/builders"
+import { Bot } from "@/bot";
+import { SlashCommandBuilder } from "@discordjs/builders";
 import Command from "@models/command";
 import { PermissionsFrom } from "@util/array_helper";
-import GetMember from "@util/member_helper";
-import { Bot } from "@/bot";
+import { CommandInteraction } from "discord.js";
 
 export default class Unban implements Command {
     name = "Unban";
     description = "Unbans a user from the guild";
-    syntax = "unban <user | userid> <reason>?";
-    aliases? = ["uban"];
 
     perms = PermissionsFrom("BAN_MEMBERS", "USE_APPLICATION_COMMANDS");
     guildOnly = true;
-
-    execute(message: Message, args: string[], BOT: Bot): void {
-        const member = GetMember(message.guild, args[0]);
-        args.shift();
-        const reason = args.join(' ') || "No reason provided";
-
-        if (!member)
-            return void(message.reply("That user doesn't exist!"));
-
-        message.guild.bans.fetch(member)
-            .then(() =>
-                message.guild.members.unban(member, reason)
-                    .then(u => message.reply(`Unbanned \`${u.tag}\` with reason: \`${reason}\``))
-                    .catch(() => message.reply("An unexpected error occured, please try again.")))
-            .catch(() => message.reply("That user is not banned!"));
-    }
 
     SlashCommand = new SlashCommandBuilder()
         .setName(this.name.toLowerCase())
@@ -39,9 +20,9 @@ export default class Unban implements Command {
     executeSlash(interaction: CommandInteraction, BOT: Bot): void {
         const user = interaction.options.getUser("user", true);
         interaction.guild.members.fetch(user).then(member => {
-            
+
             if (!member)
-                return void(interaction.reply({ content: "That user doesn't exist!", ephemeral: true }));
+                return void (interaction.reply({ content: "That user doesn't exist!", ephemeral: true }));
 
             const reason = interaction.options.getString("reason", false) || "No reason provided";
             interaction.guild.bans.fetch(member)