+++ src/commands/util/help.ts
@@ -1,37 +1,15 @@
-import { CommandInteraction, EmbedField, HexColorString, Message, MessageEmbed } from "discord.js";
+import { Bot } from "@/bot";
 import { SlashCommandBuilder } from "@discordjs/builders";
 import Command from "@models/command";
 import { PermissionsFrom } from "@util/array_helper";
-import { Bot } from "@/bot";
+import { CommandInteraction, EmbedField, HexColorString, MessageEmbed } from "discord.js";
 
 export default class Help implements Command {
     name = "Help";
     description = "Gets help for a specific command or for the bot in general";
-    syntax = "help <command>?";
-    aliases = ["?"];
 
     perms = PermissionsFrom("USE_APPLICATION_COMMANDS");
     guildOnly = false;
-
-    execute(message: Message, args: string[], BOT: Bot): void {
-        let cmd: Command = null; // Command the user passed or null if no command was specified
-
-        // Load the command if the user specified one for help
-        if (args.length > 0)
-            cmd = BOT.commands.has(args[0].toLowerCase()) ? BOT.commands.get(args[0]) : null;
-
-        // No command was specified, show list of all commands
-        if (cmd == null) {
-            message.channel.send({
-                embeds: [ this.getGeneralEmbed(message.channel.type == "DM" ? 0x8c3d1e : message.guild.me.displayHexColor, BOT) ]
-            });
-        // Show details for specified command
-        } else {
-            message.channel.send({
-                embeds: [ this.getCommandEmbed(cmd, message.channel.type == "DM" ? 0x8c3d1e : message.guild.me.displayHexColor) ]
-            });
-        }
-    }
 
     /// TODO: Add choices so the user can only select a valid command
     SlashCommand = new SlashCommandBuilder()
@@ -44,12 +22,12 @@ export default class Help implements Command {
 
         if (cmd == null) {
             interaction.reply({
-                embeds: [ this.getGeneralEmbed(!interaction.guild ? 0x8c3d1e : interaction.guild.me.displayHexColor, BOT) ],
+                embeds: [this.getGeneralEmbed(!interaction.guild ? 0x8c3d1e : interaction.guild.me.displayHexColor, BOT)],
                 ephemeral: true
             });
         } else {
             interaction.reply({
-                embeds: [ this.getCommandEmbed(cmd, !interaction.guild ? 0x8c3d1e : interaction.guild.me.displayHexColor) ],
+                embeds: [this.getCommandEmbed(cmd, !interaction.guild ? 0x8c3d1e : interaction.guild.me.displayHexColor)],
                 ephemeral: true
             });
         }
@@ -72,7 +50,7 @@ export default class Help implements Command {
             .setTimestamp()
             .addFields(fields);
     }
-    
+
     getCommandEmbed = (cmd: Command, color: number | HexColorString) => {
         return new MessageEmbed()
             .setColor(color)
@@ -80,8 +58,6 @@ export default class Help implements Command {
             .setDescription(cmd.description)
             .setTimestamp()
             .addFields(
-                { name: "Aliases", value: cmd.aliases ? cmd.aliases.map(a => `\`${a.toString().toLowerCase()}\``).join(", ") : "none", inline: false },
-                { name: "Syntax", value: cmd.syntax, inline: false },
                 { name: "Permissions", value: cmd.perms.map(p => `\`${p.toArray()}\``).join(", "), inline: false },
                 { name: "Guild Only", value: cmd.guildOnly ? "true" : "false", inline: false }
             );